Social media icons hover
Removed the hover effect of loading the new image. After some research I found that the social icons disappear on initial hover because the website needs to load this new asset onto the screen. For this reason I removed the image hover effect and increased the scale effect slightly so that the user could easily tell that the link is clickable.