contact page empty space
Added new breakpoints for the margin top of main. This is so that there is more of a balanced space between the top and bottom so that the footer is being pushed to the bottom for no reason. The contact page is something that I will be wanting to move onto the home page for a later iteration of my portfolio.