Slick slider images load
Added the script tag in the slider.js file to force load the images when the window loads to load the images. This resolved the issue of the images not loading. Also removed some of the unnecessary images in the file that did not add anything to the project. Also corrected an image that was incorrect on the AH-Photography page.